moved config struct to C file
